extract a runtime-status-icon component
