+++ ui/src/app/components/help-sidebar-icons.tsx
@@ -11,7 +11,6 @@ import {
 } from '@fortawesome/free-solid-svg-icons';
 import { faCircle } from '@fortawesome/free-solid-svg-icons/faCircle';
 import { faDna } from '@fortawesome/free-solid-svg-icons/faDna';
-import { faLock } from '@fortawesome/free-solid-svg-icons/faLock';
 import { faSyncAlt } from '@fortawesome/free-solid-svg-icons/faSyncAlt';
 import { faTerminal } from '@fortawesome/free-solid-svg-icons/faTerminal';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
@@ -20,7 +19,6 @@ import {
   CdrVersionTiersResponse,
   Criteria,
   GenomicExtractionJob,
-  RuntimeStatus,
   TerraJobStatus,
 } from 'generated/fetch';
 
@@ -28,32 +26,23 @@ import colors, { colorWithWhiteness } from 'app/styles/colors';
 import { DEFAULT, reactStyles, switchCase } from 'app/utils';
 import { getCdrVersion } from 'app/utils/cdr-versions';
 import { ComputeSecuritySuspendedError } from 'app/utils/runtime-utils';
-import {
-  CompoundRuntimeOpStore,
-  RuntimeStore,
-  serverConfigStore,
-} from 'app/utils/stores';
+import { RuntimeStore, serverConfigStore } from 'app/utils/stores';
 import { WorkspaceData } from 'app/utils/workspace-data';
 import { WorkspacePermissionsUtil } from 'app/utils/workspace-permissions';
 import { supportUrls } from 'app/utils/zendesk';
-import arrowLeft from 'assets/icons/arrow-left-regular.svg';
-import runtime from 'assets/icons/thunderstorm-solid.svg';
-import times from 'assets/icons/times-light.svg';
+import thunderstorm from 'assets/icons/thunderstorm-solid.svg';
 import moment from 'moment/moment';
 
 import { RouteLink } from './app-router';
 import { FlexRow } from './flex';
 import { TooltipTrigger } from './popups';
+import { RuntimeStatusIcon } from './runtime-status-icon';
 
 const styles = reactStyles({
   asyncOperationStatusIcon: {
     width: '.5rem',
     height: '.5rem',
     zIndex: 2,
-  },
-  runtimeStatusIconOutline: {
-    border: `1px solid ${colors.white}`,
-    borderRadius: '.25rem',
   },
   statusIconContainer: {
     alignSelf: 'flex-end',
@@ -120,32 +109,11 @@ export interface IconConfig {
   hasContent: boolean;
 }
 
-export const proIcons = {
-  arrowLeft: arrowLeft,
-  runtime: runtime,
-  times: times,
-};
-
 const displayRuntimeIcon = (
   icon: IconConfig,
-  workspace: WorkspaceData,
-  store: RuntimeStore,
-  compoundRuntimeOps: CompoundRuntimeOpStore
+  workspaceNamespace: string,
+  store: RuntimeStore
 ) => {
-  let status = store?.runtime?.status;
-  if (
-    (!status || status === RuntimeStatus.Deleted) &&
-    workspace.namespace in compoundRuntimeOps
-  ) {
-    // If a compound operation is still pending, and we're transitioning
-    // through the "Deleted" phase of the runtime, we want to keep showing
-    // an activity spinner. Avoids an awkward UX during a delete/create cycle.
-    // There also be some lag during the runtime creation flow between when
-    // the compound operation starts, and the runtime is set in the store; for
-    // this reason use Creating rather than Deleting here.
-    status = RuntimeStatus.Creating;
-  }
-
   // We always want to show the thunderstorm icon.
   // For most runtime statuses (Deleting and Unknown currently excepted), we will show a small
   // overlay icon in the bottom right of the tab showing the runtime status.
@@ -159,92 +127,10 @@ const displayRuntimeIcon = (
     >
       <img
         data-test-id={'help-sidebar-icon-' + icon.id}
-        src={proIcons[icon.id]}
+        src={thunderstorm}
         style={{ ...icon.style, position: 'absolute' }}
       />
-      <FlexRow
-        data-test-id='runtime-status-icon-container'
-        style={styles.statusIconContainer}
-      >
-        {(() => {
-          const errIcon = (
-            <FontAwesomeIcon
-              icon={faCircle}
-              style={{
-                ...styles.asyncOperationStatusIcon,
-                ...styles.runtimeStatusIconOutline,
-                color: colors.asyncOperationStatus.error,
-              }}
-            />
-          );
-
-          if (store.loadingError) {
-            if (store.loadingError instanceof ComputeSecuritySuspendedError) {
-              return (
-                <FontAwesomeIcon
-                  icon={faLock}
-                  style={{
-                    ...styles.asyncOperationStatusIcon,
-                    color: colors.asyncOperationStatus.stopped,
-                  }}
-                />
-              );
-            }
-            return errIcon;
-          }
-          switch (status) {
-            case RuntimeStatus.Creating:
-            case RuntimeStatus.Starting:
-            case RuntimeStatus.Updating:
-              return (
-                <FontAwesomeIcon
-                  icon={faSyncAlt}
-                  style={{
-                    ...styles.asyncOperationStatusIcon,
-                    ...styles.rotate,
-                    color: colors.asyncOperationStatus.starting,
-                  }}
-                />
-              );
-            case RuntimeStatus.Stopped:
-              return (
-                <FontAwesomeIcon
-                  icon={faCircle}
-                  style={{
-                    ...styles.asyncOperationStatusIcon,
-                    ...styles.runtimeStatusIconOutline,
-                    color: colors.asyncOperationStatus.stopped,
-                  }}
-                />
-              );
-            case RuntimeStatus.Running:
-              return (
-                <FontAwesomeIcon
-                  icon={faCircle}
-                  style={{
-                    ...styles.asyncOperationStatusIcon,
-                    ...styles.runtimeStatusIconOutline,
-                    color: colors.asyncOperationStatus.running,
-                  }}
-                />
-              );
-            case RuntimeStatus.Stopping:
-            case RuntimeStatus.Deleting:
-              return (
-                <FontAwesomeIcon
-                  icon={faSyncAlt}
-                  style={{
-                    ...styles.asyncOperationStatusIcon,
-                    ...styles.rotate,
-                    color: colors.asyncOperationStatus.stopping,
-                  }}
-                />
-              );
-            case RuntimeStatus.Error:
-              return errIcon;
-          }
-        })()}
-      </FlexRow>
+      <RuntimeStatusIcon {...{ store, workspaceNamespace }} />
     </FlexRow>
   );
 };
@@ -408,18 +294,11 @@ interface DisplayIconProps {
   criteria: Array<Selection>;
   concept?: Array<Criteria>;
   runtimeStore: RuntimeStore;
-  compoundRuntimeOps: CompoundRuntimeOpStore;
   genomicExtractionJobs: GenomicExtractionJob[];
 }
 const displayIcon = (icon: IconConfig, props: DisplayIconProps) => {
-  const {
-    workspace,
-    runtimeStore,
-    compoundRuntimeOps,
-    genomicExtractionJobs,
-    criteria,
-    concept,
-  } = props;
+  const { workspace, runtimeStore, genomicExtractionJobs, criteria, concept } =
+    props;
   return switchCase(
     icon.id,
     [
@@ -436,8 +315,7 @@ const displayIcon = (icon: IconConfig, props: DisplayIconProps) => {
     ],
     [
       'runtime',
-      () =>
-        displayRuntimeIcon(icon, workspace, runtimeStore, compoundRuntimeOps),
+      () => displayRuntimeIcon(icon, workspace.namespace, runtimeStore),
     ],
     [
       'terminal',
@@ -464,7 +342,7 @@ const displayIcon = (icon: IconConfig, props: DisplayIconProps) => {
         icon.faIcon === null ? (
           <img
             data-test-id={'help-sidebar-icon-' + icon.id}
-            src={proIcons[icon.id]}
+            src={icon.id === 'runtime' && thunderstorm}
             style={icon.style}
           />
         ) : (
@@ -612,7 +490,6 @@ const iconConfig = (
 interface HelpSidebarIconsProps extends IconConfigProps {
   workspace: WorkspaceData;
   cdrVersionTiersResponse: CdrVersionTiersResponse;
-  compoundRuntimeOps: CompoundRuntimeOpStore;
   genomicExtractionJobs: GenomicExtractionJob[];
   activeIcon: string;
   onIconClick: (icon: IconConfig) => void;
